+++ test/unit/mongo_rep_admin/rep_msg_chk.py
@@ -17,7 +17,6 @@
 import sys
 import os
 import unittest
-import mock
 
 # Local
 sys.path.append(os.getcwd())
@@ -35,8 +34,10 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp
-        test_no_error_msg
-        test_error_msg
+        test_error_msg_true_nr_false
+        test_error_msg_true_nr_true
+        test_error_msg_false_nr_false
+        test_error_msg_false_nr_true
 
     """
 
@@ -50,35 +51,71 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        self.rep_stat = {"infoMessage": "StringHere"}
-        self.rep_stat2 = {"state": 3, "stateStr": "StringHere"}
+        self.rep_stat = {"infoMessage": None}
+        self.rep_stat2 = {"infoMessage": "StringHere"}
+#        self.rep_stat2 = {"state": 3, "stateStr": "StringHere"}
+        self.server = {}
+        self.no_report = True
+        self.no_report2 = False
+        self.results = {}
+        self.results2 = {'ErrorMessage': None}
+        self.results3 = {'ErrorMessage': 'StringHere'}
 
-    def test_no_error_msg(self):
+    def test_error_msg_true_nr_false(self):
 
-        """Function:  test_no_error_msg
+        """Function:  test_error_msg_true_nr_false
 
-        Description:  Test with no error message.
-
-        Arguments:
-
-        """
-
-        self.assertFalse(mongo_rep_admin.rep_msg_chk(self.rep_stat2))
-
-    @mock.patch("mongo_rep_admin.gen_libs.prt_msg")
-    def test_error_msg(self, mock_prt):
-
-        """Function:  test_error_msg
-
-        Description:  Test with error message.
+        Description:  Test with error message false and no report false.
 
         Arguments:
 
         """
 
-        mock_prt.return_value = True
+        self.assertEqual(
+            mongo_rep_admin.rep_msg_chk(
+                self.rep_stat2, self.server, self.no_report2), self.results3)
 
-        self.assertFalse(mongo_rep_admin.rep_msg_chk(self.rep_stat))
+    def test_error_msg_true_nr_true(self):
+
+        """Function:  test_error_msg_true_nr_true
+
+        Description:  Test with error message false and no report true.
+
+        Arguments:
+
+        """
+
+        self.assertEqual(
+            mongo_rep_admin.rep_msg_chk(
+                self.rep_stat2, self.server, self.no_report), self.results3)
+
+    def test_error_msg_false_nr_false(self):
+
+        """Function:  test_error_msg_false_nr_false
+
+        Description:  Test with error message false and no report false.
+
+        Arguments:
+
+        """
+
+        self.assertEqual(
+            mongo_rep_admin.rep_msg_chk(
+                self.rep_stat, self.server, self.no_report2), self.results2)
+
+    def test_error_msg_false_nr_true(self):
+
+        """Function:  test_error_msg_false_nr_true
+
+        Description:  Test with error message false and no report true.
+
+        Arguments:
+
+        """
+
+        self.assertEqual(
+            mongo_rep_admin.rep_msg_chk(
+                self.rep_stat, self.server, self.no_report), self.results)
 
 
 if __name__ == "__main__":