+++ CHANGELOG.md
@@ -19,6 +19,8 @@ Breaking changes
 - create_data_config: Create data out config parameters.
 
 ### Changed
+- chk_rep_stat: Refactored function to remove prints and replace with dictionary and call mongo_libs.data_out to handle data out function.
+- rep_msg_chk, rep_health_chk, rep_state_chk: Removed prints and replaced with dictionary and returned dictionary to calling function, modied input arguments.
 - node_chk: Refactored function to remove any processing of arguments, replaced data out commands with call to mongo_libs.data_out and added pre-header to JSON output.
 - chk_mem_rep_lag: Refactored function to remove any processing of arguments, replaced datetime with TimeFormat class instance, removed process_std and process_json calls, replaced with mongo_libs.data_out call and replaced JSON header creation with call to create_header.
 - chk_rep_lag: Refactored function to remove any processing of arguments contained in data out configuration.